GhostSec/requirements.txt
psutil>=5.9.0
requests>=2.28.0
pytest>=8.4.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0